pyproject.toml
[build-system]
requires = ["setuptools>=63.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "unsim"
description = "UNsim: Differentiable network traffic simulation in Python"
readme = "README.md"
authors = [{name = "Toru Seo", email = "seo.t.aa@m.titech.ac.jp"}]
license = "MIT"
classifiers = [
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering",
]
requires-python = ">=3.10"
dependencies = ["numpy", "pandas", "matplotlib", "pillow", "pytest"] #jax must be installed by user as the optimal version depends on the hard/software env
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/toruseo/UNsim"


[tool.setuptools]
packages = {find = {include = ["unsim", "unsim.*"]}}

[tool.setuptools.dynamic]
version = {attr = "unsim.__version__"}